refactors to share code between react and expo
