+++ apps/wallets/shared/package.json
@@ -1,0 +1,7 @@
+{
+    "name": "@crossmint/wallets-playground-shared",
+    "version": "0.0.1",
+    "private": true,
+    "main": "./src/index.ts",
+    "types": "./src/index.ts"
+}
