+++ apps/wallets/shared/src/constants/tokens.ts
@@ -1,0 +1,2 @@
+export const TOKENS = ["usdxm", "usdc"] as const;
+export type Token = (typeof TOKENS)[number];
