+++ apps/wallets/shared/src/types/signer.ts
@@ -1,0 +1,2 @@
+export const SIGNER_TYPES = ["device", "passkey", "external-wallet"] as const;
+export type SignerType = (typeof SIGNER_TYPES)[number];
