+++ apps/wallets/shared/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../../../tsconfig.base.json",
+    "compilerOptions": {
+        "rootDir": "src",
+        "jsx": "react-jsx"
+    },
+    "include": ["src/**/*"]
+}
